+++ src/ska_tmc_sdpmasterleafnode/commands/__init__.py
@@ -3,6 +3,9 @@
 from ska_tmc_sdpmasterleafnode.commands.disable_command import Disable
 from ska_tmc_sdpmasterleafnode.commands.off_command import Off
 from ska_tmc_sdpmasterleafnode.commands.on_command import On
+from ska_tmc_sdpmasterleafnode.commands.set_controller_admin_mode import (
+    SetAdminMode,
+)
 from ska_tmc_sdpmasterleafnode.commands.standby_command import Standby
 
-__all__ = ["Disable", "Off", "On", "Standby"]
+__all__ = ["Disable", "Off", "On", "Standby", "SetAdminMode"]