+++ src/ska_tmc_sdpsubarrayleafnode/commands/sdp_sln_command.py
@@ -9,7 +9,7 @@ from ska_control_model.task_status import TaskStatus
 from ska_ser_logging import configure_logging
 from ska_tango_base.base import TaskCallbackType
 from ska_tango_base.commands import ResultCode
-from ska_tmc_common import SdpSubArrayAdapter, TimeKeeper
+from ska_tmc_common import SdpSubArrayAdapter
 from ska_tmc_common.adapters import AdapterType
 from ska_tmc_common.exceptions import CommandNotAllowed
 from ska_tmc_common.tmc_command import TmcLeafNodeCommand
@@ -62,9 +62,6 @@ class SdpSLNCommand(TmcLeafNodeCommand):
         self.component_manager = component_manager
         self.sdp_subarray_adapter = None
         self.task_callback: TaskCallbackType = task_callback_default
-        self.timekeeper = TimeKeeper(
-            self.component_manager.command_timeout, logger
-        )
 
     def check_op_state(self, command_name) -> None:
         """Checks the operational state of the device"""
@@ -83,11 +80,14 @@ class SdpSLNCommand(TmcLeafNodeCommand):
             )
 
     def init_adapter(self) -> Tuple[ResultCode, str]:
-        timeout = self.component_manager.timeout
+        adapter_timeout = self.component_manager.adapter_timeout
         elapsed_time: float = 0
         start_time: float = time.time()
         device = self.component_manager._sdp_subarray_dev_name
-        while self.sdp_subarray_adapter is None and elapsed_time < timeout:
+        while (
+            self.sdp_subarray_adapter is None
+            and elapsed_time < adapter_timeout
+        ):
             try:
                 get_adapter = self.adapter_factory.get_or_create_adapter
                 self.sdp_subarray_adapter: SdpSubArrayAdapter = get_adapter(
@@ -96,7 +96,7 @@ class SdpSLNCommand(TmcLeafNodeCommand):
                 )
             except ConnectionFailed as connection_failed:
                 elapsed_time = time.time() - start_time
-                if elapsed_time > timeout:
+                if elapsed_time > adapter_timeout:
                     message = (
                         "Error in creating adapter for %s : %s",
                         device,
@@ -105,7 +105,7 @@ class SdpSLNCommand(TmcLeafNodeCommand):
                     return ResultCode.FAILED, message
             except DevFailed as device_failed:
                 elapsed_time = time.time() - start_time
-                if elapsed_time > timeout:
+                if elapsed_time > adapter_timeout:
                     message = (
                         "Error in creating adapter for %s : %s",
                         device,