+++ src/ska_tmc_sdpsubarrayleafnode/manager/event_receiver.py
@@ -8,7 +8,7 @@ import tango
 from ska_ser_logging import configure_logging
 from ska_tango_base.control_model import ObsState
 from ska_tmc_common.device_info import SubArrayDeviceInfo
-from ska_tmc_common.event_receiver import EventReceiver
+from ska_tmc_common.v1.event_receiver import EventReceiver
 
 configure_logging()
 LOGGER: logging.Logger = logging.getLogger(__name__)
@@ -31,17 +31,17 @@ class SdpSLNEventReceiver(EventReceiver):
         logger: logging.Logger = LOGGER,
         max_workers: int = 1,
         proxy_timeout: int = 500,
-        sleep_time: int = 1,
+        event_subscription_check_period: int = 1,
     ):
         super().__init__(
             component_manager=component_manager,
             logger=logger,
             max_workers=max_workers,
             proxy_timeout=proxy_timeout,
-            sleep_time=sleep_time,
+            event_subscription_check_period=event_subscription_check_period,
         )
         self._max_workers = max_workers
-        self._sleep_time = sleep_time
+        self._event_subscription_check_period = event_subscription_check_period
         self._stop = False
         self._component_manager = component_manager
 
@@ -53,27 +53,39 @@ class SdpSLNEventReceiver(EventReceiver):
                 dev_info = self._component_manager.get_device()
                 if dev_info.last_event_arrived is None:
                     executor.submit(self.subscribe_events, dev_info)
-            sleep(self._sleep_time)
+            sleep(self._event_subscription_check_period)
 
     def subscribe_events(
         self,
         dev_info: SubArrayDeviceInfo,
         attribute_dictionary: Optional[Dict[str, str]] = None,
     ):
+        sdp_subarray_proxy = None
         try:
-            proxy = self._dev_factory.get_device(dev_info.dev_name)
-            proxy.subscribe_event(
+            sdp_subarray_proxy = self._dev_factory.get_device(
+                dev_info.dev_name
+            )
+            sdp_subarray_proxy.subscribe_event(
                 "obsState",
                 tango.EventType.CHANGE_EVENT,
                 self.handle_obs_state_event,
                 stateless=True,
             )
-        except Exception as exception:
-            self._logger.debug(
-                "Error on Device %s while event subscription. Exception: %s",
-                proxy.dev_name,
-                exception,
+            sdp_subarray_proxy.subscribe_event(
+                "adminMode",
+                tango.EventType.CHANGE_EVENT,
+                self.handle_admin_mode_event,
+                stateless=True,
             )
+            self.stop()
+        except Exception as exception:
+            if sdp_subarray_proxy:
+                self._logger.debug(
+                    "Error on Device %s while event subscription."
+                    " Exception: %s",
+                    sdp_subarray_proxy.dev_name,
+                    exception,
+                )
 
     def handle_obs_state_event(self, event):
         """